+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### added
 * get_bar 增加聚宽源
 * get_daily 支持基金累计净值
+* get_rt 增加时间属性
 
 ## v0.8.4 - 2020.04.06
 ### added